Rename dir, separate murmur2, make deterministic
