--- gungnir/src/stats.rs
@@ -1,2 +1,0 @@
-pub mod tdigest;
-pub mod hyperloglog;
